update python wrapper to allow calculating average, outputting to stdout
breaking changes: outputting to stdout by default; if an output file is needed,
specify the -o option.

+++ parse-qp-output.py
@@ -12,7 +12,20 @@ import sys
 import json
 
 
-def parse_file(input_file):
+def print_stderr(msg):
+    print(msg, file=sys.stderr)
+
+
+def average(x):
+    if not isinstance(x, list):
+        print_stderr("Cannot use average() on non-list!")
+        sys.exit(1)
+    if not x:
+        return []
+    return sum(x) / len(x)
+
+
+def parse_file(input_file, use_average=False):
     with open(input_file) as f:
         frame_index = -1
         first_frame_found = False
@@ -39,14 +52,14 @@ def parse_file(input_file):
                     yield {
                         "frameType": frame_type,
                         "frameSize": frame_size,
-                        "qpValues": frame_qp_values
+                        "qpValues": frame_qp_values if not use_average else [average(frame_qp_values)]
                     }
 
                 first_frame_found = True
 
                 frame_type = line[-1]
                 if frame_type not in ["I", "P", "B"]:
-                    print("Wrong frame type parsed: " + str(frame_type))
+                    print_stderr("Wrong frame type parsed: " + str(frame_type))
                     sys.exit(1)
                 frame_index += 1
                 # initialize empty for the moment
@@ -75,18 +88,18 @@ def parse_file(input_file):
                 # remove the leading space in case of single digit qp values
                 line_qp_values = [int(raw_values[i:i + 2].lstrip())
                              for i in range(0, len(raw_values), 2)]
-                # print("Adding QP values to frame with index " + str(frame_index))
+                # print_stderr("Adding QP values to frame with index " + str(frame_index))
                 frame_qp_values.extend(line_qp_values)
                 continue
             if "pkt_size" in line:
-                frame_size = re.findall(r'\d+', line)[0]
+                frame_size = int(re.findall(r'\d+', line)[0])
 
         # yield last frame
         if has_current_frame_data:
             yield {
                 "frameType": frame_type,
                 "frameSize": frame_size,
-                "qpValues": frame_qp_values
+                "qpValues": frame_qp_values if not use_average else [average(frame_qp_values)]
             }
 
 
@@ -95,24 +108,29 @@ def main():
     parser = argparse.ArgumentParser(
         description="Parse QP values from ffmpeg-debug-qp")
     parser.add_argument("input", type=str, help="Input log file")
-    parser.add_argument("output", type=str, help="Output LD-JSON file")
+    parser.add_argument("-o", "--output", help="Output LD-JSON file")
     parser.add_argument("-f", "--force", action="store_true", help="Overwrite output")
+    parser.add_argument("-a", "--average", action="store_true", help="Calculate only average QP (efficiency)")
 
     args = vars(parser.parse_args())
     if not os.path.isfile(args["input"]):
-        print("no such file: " + args["input"])
+        print_stderr("no such file: " + args["input"])
         sys.exit(1)
 
-    if os.path.isfile(args["output"]) and not args["force"]:
-        print("output " + args["output"] + " already exists; use -f/--force to overwrite")
-        sys.exit(1)
+    if args["output"]:
+        if os.path.isfile(args["output"]) and not args["force"]:
+            print_stderr("output " + args["output"] + " already exists; use -f/--force to overwrite")
+            sys.exit(1)
 
-    with open(args["output"], 'w'):
-        pass
-    with open(args["output"], "a") as of:
-        of.truncate()
-        for data in parse_file(args["input"]):
-            of.write(json.dumps(data) + "\n")
+        with open(args["output"], 'w'):
+            pass
+        with open(args["output"], "a") as of:
+            of.truncate()
+            for data in parse_file(args["input"], args["average"]):
+                of.write(json.dumps(data) + "\n")
+    else:
+        for data in parse_file(args["input"], args["average"]):
+            print(json.dumps(data))
 
 
 if __name__ == '__main__':
